public file added to backend commit
as a way to clean up all the file works I've decided to keep a single files system which hold a public folder within the backend code works to persist a neutral functionality over an unnecessary call function with such a small project.
this update include added features such as --- register user, update/delete recipes and most of all a persistent login score so when a user is to reload their creds stay featured with the login up until the user decides when to logout.